pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["ClearClown(清楚系ピエロ) <clearcleaclown@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.34"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
